+++ src/main/gen/com/en_circle/slt/plugin/lisp/psi/LispEnhancement.java
@@ -2,7 +2,11 @@
 package com.en_circle.slt.plugin.lisp.psi;
 
 import com.intellij.psi.PsiElement;
+import org.jetbrains.annotations.Nullable;
 
 public interface LispEnhancement extends PsiElement {
+
+  @Nullable
+  LispFunctionEnhancement getFunctionEnhancement();
 
 }