+++ src/main/java/com/en_circle/slt/plugin/actions/MacroexpandActionBase.java
@@ -70,7 +70,7 @@ public abstract class MacroexpandActionBase extends AnAction {
                     JBPopup popup = JBPopupFactory.getInstance()
                             .createComponentPopupBuilder(textPanel, null)
                             .setProject(editor.getProject())
-                            .setTitle(SltBundle.message("slt.documentation.macroexpand"))
+                            .setTitle(SltBundle.message("slt.ui.macroexpand.title"))
                             .setShowBorder(true)
                             .setMovable(true)
                             .setFocusable(true)