+++ src/main/java/com/en_circle/slt/plugin/actions/swank/SltSwankAction.java
@@ -1,0 +1,37 @@
+package com.en_circle.slt.plugin.actions.swank;
+
+import com.en_circle.slt.plugin.SltCommonLispFileType;
+import com.intellij.openapi.actionSystem.ActionUpdateThread;
+import com.intellij.openapi.actionSystem.AnAction;
+import com.intellij.openapi.actionSystem.AnActionEvent;
+import com.intellij.openapi.actionSystem.CommonDataKeys;
+import com.intellij.openapi.editor.Editor;
+import com.intellij.psi.PsiDocumentManager;
+import com.intellij.psi.PsiFile;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.Objects;
+
+public abstract class SltSwankAction extends AnAction {
+
+    protected abstract boolean innerActionCheck(Editor editor, PsiFile file);
+
+    @Override
+    public @NotNull ActionUpdateThread getActionUpdateThread() {
+        return ActionUpdateThread.BGT;
+    }
+
+    @Override
+    public void update(@NotNull AnActionEvent event) {
+        event.getPresentation().setEnabledAndVisible(false);
+
+        Editor editor = event.getData(CommonDataKeys.EDITOR);
+        if (editor != null && event.getProject() != null) {
+            PsiFile file = PsiDocumentManager.getInstance(Objects.requireNonNull(editor.getProject())).getPsiFile(editor.getDocument());
+            if (file != null && SltCommonLispFileType.INSTANCE.equals(file.getFileType())) {
+                event.getPresentation().setEnabledAndVisible(innerActionCheck(editor, file));
+            }
+        }
+    }
+
+}