+++ src/main/java/com/en_circle/slt/plugin/services/lisp/components/SltIndentationContainer.java
@@ -188,7 +188,11 @@ public class SltIndentationContainer {
         if (element.getParent() != file) {
             // we are in correct form
             LispToplevel toplevel = PsiTreeUtil.getParentOfType(element, LispToplevel.class);
-            assert toplevel != null;
+            if (toplevel == null) {
+                // something very wrong, like bad lisp code, do not indent
+                return 0;
+            }
+
             int numBraces = 0;
 
             LispList parent = PsiTreeUtil.getParentOfType(element, LispList.class);
@@ -198,8 +202,8 @@ public class SltIndentationContainer {
             }
 
             try {
-                String formText = documentText.substring(toplevel.getTextOffset(), offset);
-                boolean isUnfinished = false;
+                String formText = documentText.substring(toplevel.getTextOffset(),
+                        element.getTextOffset() + element.getTextLength() - (wasAfter ? 0 : 1));
                 LispList list = PsiTreeUtil.getParentOfType(element, LispList.class);
                 if (list != null) {
                     if (element.getNode().getElementType() == LispTypes.RPAREN) {
@@ -207,7 +211,6 @@ public class SltIndentationContainer {
                     }
                     if (list != null) {
                         if (list.getNextSibling() instanceof PsiErrorElement) {
-                            isUnfinished = true;
                             if (element.getNode().getElementType() == LispTypes.RPAREN) {
                                 if (numBraces > 0) {
                                     --numBraces;
@@ -216,11 +219,8 @@ public class SltIndentationContainer {
                         }
                     }
                 }
-                if (!wasAfter || isUnfinished) {
-                    // insert fake element at the end, so we identify correct form
-                    state.hasRealElement = false;
-                    formText += " 0";
-                }
+                state.hasRealElement = false;
+                formText += " 0";
                 formText += StringUtils.repeat(')', numBraces);
                 return calculateIndent(state, formText, file.getProject());
             } catch (Exception ignored) {
@@ -426,7 +426,7 @@ public class SltIndentationContainer {
         if (indentation != null) {
             // we are not backtracking and we found the main rule
 
-            // since we actually matched head we consider this list "toplevel" with regards to indent
+            // since we actually matched head we consider this list "toplevel" in regard to indent
             topLevel = container;
 
             if (indentation.normalArgumentCount != null) {