+++ src/main/java/com/en_circle/slt/plugin/swank/SwankPacket.java
@@ -388,6 +388,46 @@ public class SwankPacket {
         return new SwankPacket(formatted);
     }
 
+    public static SwankPacket disassemble(String symbol, String packageName, BigInteger continuation) {
+        symbol = StringUtils.replace(symbol, "\\", "\\\\");
+        symbol = StringUtils.replace(symbol, "\"", "\\\"");
+//        packageName = StringUtils.replace(packageName, "\\", "\\\\");
+//        packageName = StringUtils.replace(packageName, "\"", "\\\"");
+
+        String formatted = String.format("(:emacs-rex (swank:disassemble-form \"'%s\") \":CL-USER\" T %s)", symbol, continuation);
+        return new SwankPacket(formatted);
+    }
+
+    public static SwankPacket undefineFunction(String symbol, String packageName, BigInteger continuation) {
+        symbol = StringUtils.replace(symbol, "\\", "\\\\");
+        symbol = StringUtils.replace(symbol, "\"", "\\\"");
+//        packageName = StringUtils.replace(packageName, "\\", "\\\\");
+//        packageName = StringUtils.replace(packageName, "\"", "\\\"");
+
+        String formatted = String.format("(:emacs-rex (swank:undefine-function \"%s\") \":CL-USER\" T %s)", symbol, continuation);
+        return new SwankPacket(formatted);
+    }
+
+    public static SwankPacket uninternSymbol(String symbol, String packageName, BigInteger continuation) {
+        symbol = StringUtils.replace(symbol, "\\", "\\\\");
+        symbol = StringUtils.replace(symbol, "\"", "\\\"");
+        packageName = StringUtils.replace(packageName, "\\", "\\\\");
+        packageName = StringUtils.replace(packageName, "\"", "\\\"");
+
+        String formatted = String.format("(:emacs-rex (swank:unintern-symbol \"%s\" \"%s\") \":CL-USER\" T %s)", symbol, packageName,
+                continuation);
+        return new SwankPacket(formatted);
+    }
+
+    public static SwankPacket reevaluteDefvar(String form, BigInteger continuation) {
+        form = StringUtils.replace(form, "\\", "\\\\");
+        form = StringUtils.replace(form, "\"", "\\\"");
+
+        String formatted = String.format("(:emacs-rex (swank:re-evaluate-defvar \"%s\" ) \":CL-USER\" T %s)", form,
+                continuation);
+        return new SwankPacket(formatted);
+    }
+
     private int length;
     private String expressionSource;